+++ app/controllers/comparisons/change_in_electricity_consumption_recent_school_weeks_controller.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
 module Comparisons
-  class ChangeInElectricityConsumptionRecentSchoolWeeksController < BaseController
-    include ChangeInConsumption
-
+  class ChangeInElectricityConsumptionRecentSchoolWeeksController < Shared::ChangeInConsumption
     private
 
     def headers
