+++ app/controllers/comparisons/change_in_gas_consumption_recent_school_weeks_controller.rb
@@ -1,7 +1,5 @@
 module Comparisons
-  class ChangeInGasConsumptionRecentSchoolWeeksController < BaseController
-    include ChangeInConsumption
-
+  class ChangeInGasConsumptionRecentSchoolWeeksController < Shared::ChangeInConsumption
     private
 
     def headers
