+++ app/controllers/comparisons/change_in_gas_holiday_consumption_previous_holiday_controller.rb
@@ -1,7 +1,5 @@
 module Comparisons
-  class ChangeInGasHolidayConsumptionPreviousHolidayController < BaseController
-    include ChangeInConsumption
-
+  class ChangeInGasHolidayConsumptionPreviousHolidayController < Shared::ChangeInConsumption
     private
 
     def key
