+++ app/controllers/comparisons/change_in_gas_holiday_consumption_previous_years_holiday_controller.rb
@@ -1,7 +1,5 @@
 module Comparisons
-  class ChangeInGasHolidayConsumptionPreviousYearsHolidayController < BaseController
-    include ChangeInConsumption
-
+  class ChangeInGasHolidayConsumptionPreviousYearsHolidayController < Shared::ChangeInConsumption
     private
 
     def key
