+++ app/controllers/comparisons/gas_consumption_during_holiday_controller.rb
@@ -1,7 +1,5 @@
 module Comparisons
-  class GasConsumptionDuringHolidayController < BaseController
-    include ConsumptionDuringHoliday
-
+  class GasConsumptionDuringHolidayController < Shared::ConsumptionDuringHoliday
     private
 
     def key
