--- app/views/comparisons/change_in_electricity_consumption_recent_school_weeks/_table.html.erb
@@ -1,1 +1,0 @@
-<%= render partial: 'comparisons/shared/change_in_consumption/table', locals: local_assigns %>
