--- app/views/comparisons/change_in_electricity_holiday_consumption_previous_holiday/_table.html.erb
@@ -1,1 +1,0 @@
-<%= render partial: 'comparisons/shared/change_in_consumption/table', locals: local_assigns %>
