--- app/views/comparisons/change_in_gas_consumption_recent_school_weeks/_table.csv.ruby
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-render 'comparisons/shared/change_in_consumption/table'
