--- app/views/comparisons/change_in_gas_holiday_consumption_previous_years_holiday/_table.csv.ruby
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-render 'comparisons/shared/change_in_consumption/table'
