--- app/views/comparisons/electricity_consumption_during_holiday/_table.html.erb
@@ -1,1 +1,0 @@
-<%= render partial: 'comparisons/shared/consumption_during_holiday/table', locals: local_assigns %>
