--- app/views/comparisons/gas_consumption_during_holiday/_table.csv.ruby
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-render 'comparisons/shared/consumption_during_holiday/table'
